backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
alembic
websockets
pytest
httpx
pytest-asyncio
anthropic[bedrock]
boto3
pandas
pandas-ta
numpy
apscheduler
